backend/requirements.txt
aniso8601==8.0.0
autopep8==1.5.4
botocore==1.17.50
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
google-api-core==1.22.1
google-auth==1.21.0
google-cloud-core==1.4.1
google-cloud-datastore==1.15.0
google-cloud-ndb==1.5.0
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyMySQL==0.10.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
gunicorn==19.3.0